Makefile: Build each target by default, with a different filename
Avoid 'make HARDWARE=xyz'.
It's extremely easy to get confused and load the wrong firmware when
the build result is always the same filename. Instead, encode the
target name in the build output. This also allows us to build all
targets during one invocation of make. This allows us to kill the
'HARDWARE=' variable entirely.