Makefile: Separate rules for packihx call
The rule for "usbjtag-%.hex" would use a temporary file for the
packihx step. Because it used the same name, it often failed or
produced incorrect results with parallel make. To fix this, make pre-
and post- packihx targets, and do not use a temporary file.
This has the advantage of using make's dependency tracking, and solves
the build failures with parallel make.